ml_project/requirements.txt
matplotlib==3.5.1
pandas==1.3.5
seaborn==0.11.2
scikit-learn==1.0.2
numpy==1.21.6
marshmallow-dataclass==8.5.5
pyyaml==3.11
click==8.1.3